PR sets: create PR set selectors
spr update needs a way so selecting groups of commits that will be included in the pull request set. The selector language provides several ways of selecting commits. They can be selected by:
1. Commit ids.
2. Commit id ranges.
3. Existing PR set.

The selector language also allows users to specify a destination PR set so that existing PR sets can be updated with different commits.

The PR sets is still a WIP.

commit-id:c58f4db6